Add Dockerfile for ckb-daemon together with a docker readme.
This runs ckb-daemon in a docker container. When I start this
I can succesfully run commands against my keyboard so I assume
that this works properly. The only problem I encountered is when
I stop the container, my keyboard does not work. This problem is
listed in the known issues section in the README.md.